requirements/prod.txt
django==2.2.10
django-environ==0.4.5
django-filter==2.1.0
django-ckeditor==5.6.1
django-mptt==0.9.1
django-admin-tools==0.8.1
django-imagekit==4.0.2
django-cities-light==3.5.0
django-model-utils==3.2.0
sendgrid-django==4.2.0
djangorestframework==3.9.2
markdown==3.0.1
celery==4.2.1
kombu==4.2.2.post1
coreapi==2.3.3
psycopg2-binary==2.8.4
pillow==6.2.1
coreapi-cli==1.0.9
uwsgi==2.0.18
unidecode==1.0.23
